app/requirements.txt
flask
requests
git+git://github.com/sch3m4/pyadb@master#egg=pyadb
python-daemon
